+++ client/src/lib/forms/disable-save.js
@@ -368,7 +368,7 @@ function forceShowForm(stateData, componentProps) {
       componentProps.innerFormProps
     );
   }
-
+  console.log("AAAAA");
   return openForm;
 }