+++ unit-tests/forms/wizard.test.js
@@ -57,7 +57,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -840,7 +840,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -1622,7 +1622,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -2058,7 +2058,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -2485,7 +2485,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -2989,7 +2989,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
       },
       resource_groups: [
         { use_prefix: true, name: "service-rg", use_data: false },
@@ -3144,7 +3144,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -3640,7 +3640,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -4110,7 +4110,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -4581,7 +4581,7 @@ describe("setup wizard", () => {
         enable_classic: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
       },
       resource_groups: [
         { use_prefix: true, name: "service-rg", use_data: false },
@@ -4998,7 +4998,7 @@ describe("setup wizard", () => {
         enable_power_vs: true,
         enable_classic: false,
         power_vs_zones: ["dal10"],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -5417,7 +5417,7 @@ describe("setup wizard", () => {
         enable_power_vs: true,
         enable_classic: false,
         power_vs_zones: ["dal10"],
-        craig_version: "1.7.0",
+        craig_version: "1.7.1",
         power_vs_high_availability: false,
       },
       resource_groups: [